Update Version 8.5.6870

+++ Swift/TUIKit/TUIContact/UI_Minimalist/Cell/CellData/TUIContactConversationCellData_Minimalist.swift
@@ -39,7 +39,7 @@ class TUIContactConversationCellData_Minimalist: TUICommonCellData {
     }
 
     override func height(ofWidth width: CGFloat) -> CGFloat {
-        return TConversationCell_Height
+        return CGFloat(TConversationCell_Height)
     }
 
     override func isEqual(_ object: Any?) -> Bool {
